rename csv_operations.py to stats_operations.py
