+++ www/stats_operations.py
